requirements.txt
python-telegram-bot[job-queue]==20.7
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
openpyxl==3.1.2
python-dotenv==1.0.0
urllib3==2.0.7
requests==2.31.0
et-xmlfile==1.1.0  # openpyxl için gerekli 